# improve interpolation performance with big input range