@javache 2022-04-08
Thanks for spotting this! I think all of the sanity checks in this file should probably be disabled in release builds. That will also enable dead code stripping to remove these checks.

Can you add an `if (__DEV__) {` check around the calls to checkInfiniteRange / checkValidInputRange and other invariants inside `createInterpolation` instead?