@Almouro 2022-04-08
Hi @javache, thanks!

It feels to me like this could warrant a bigger discussion on `invariant` in release in general. What do you think?

- One pro of having them in release is they pop up on crash reporting tools such as Sentry, and can useful to analyze bugs.
- On the other hand, keeping them can be bad for performance, so if we go this route here, I'd be happy to codemod it and surround every `invariant` in RN by `if (__DEV__)`