@javache 2022-04-08
Some invariants are important to keep around in production, because they may be states we really don't want the application to get into and do want to see escalated to crash reporting.

Other invariants, like these ones, in my opinion, are more about catching developer errors, and can be skipped in release. The trade-off here is that we may miss some errors if the inputs to the interpolate call are not static, but I think that's reasonable here.