@Almouro 2022-04-08
Cool, thanks for the explanation! 🙏 
Added the `__DEV__` wrapping as well then